requirements.txt
# Python dependencies
transformers
diffusers
torch
torchvision
moviepy==1.0.3
gTTS
Pillow
openai
python-pptx
PyMuPDF
pdf2image
mediapipe
flask

#  new packages:
tf-keras
pandas
PyPDF2==3.0.1
tensorflow
scikit-learn
pyttsx3
joblib
python-docx
pydub
librosa
soundfile
PyQt5
sentence-transformers
tqdm
matplotlib
opencv-contrib-python
imageio
imageio-ffmpeg
tensorflow-cpu==2.19.0
numpy==1.26.4
scipy==1.15.3
huggingface-hub==0.32.6
tokenizers==0.21.1
safetensors==0.5.3
sounddevice==0.5.2
soxr==0.5.0.post1
audioread==3.0.1
imageio-ffmpeg==0.6.0
opencv-contrib-python==4.11.0.86
python-docx==1.1.2
python-pptx==1.0.2
XlsxWriter==3.2.3
regex==2024.11.6
python-dotenv==1.1.0
requests==2.32.4
pywin32==310
pypiwin32==223
rich==14.0.0